Add code to patch upper-half's libc
Upper-half's libc makes calls to sbrk/mmap for memory allocation. These
calls do not go through PLT and go directly to libc's internal
implementations. This allows it to escape out of the upper half and
allocate memory in the lower half, which leads to crashes on restart
due to the dangling pointers (in upper-half's libc). (Note that we don't
save and restore any of the lower-half's memory regions.)

So, this patch inserts trampolines on upper-half's libc's sbrk/mmap
functions to avoid this problem.